Renamed and reassigned menu related context and hooks
