+++ src/contexts/MailContext.js
@@ -1,5 +1,5 @@
 import React, { createContext, useState, useEffect } from 'react';
-import useAuth from '../hooks/useAuth';
+import useAuth from '../hooks/contexthooks/useAuth';
 
 export const MailContext = createContext(null);
 // Prevents access to add-card component if a card  hasen't been selected
