+++ src/contexts/SearchContext.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect, createContext } from 'react';
-import useAuth from '../hooks/useAuth';
+import useAuth from '../hooks/contexthooks/useAuth';
 export const SearchContext = createContext(null);
 
 export const SearchProvider = ({ children }) => {
