+++ src/features/mail/ComposeMessage.js
@@ -1,7 +1,7 @@
 import React, { useContext, useState, useEffect, useRef } from 'react';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { PathContext } from '../../contexts/PathContext';
-import useAuth from '../../hooks/useAuth';
+import useAuth from '../../hooks/contexthooks/useAuth';
 import validate from '../../assets/utilities/validateMessage';
 import getPath from '../../assets/utilities/getPath';
 import useMessage from '../../hooks/useMessage';
