+++ src/features/mail/Inbox.js
@@ -6,7 +6,7 @@ import Mail from './Mail';
 import ComposeMessage from './ComposeMessage';
 import Loading from '../../layout/Loading';
 import Message from './Message';
-import useAuth from '../../hooks/useAuth';
+import useAuth from '../../hooks/contexthooks/useAuth';
 import getPath from '../../assets/utilities/getPath';
 import { PathContext } from '../../contexts/PathContext';
 import { api } from '../../api/resources';
