+++ src/features/mail/Message.js
@@ -4,7 +4,7 @@ import { FiArrowLeftCircle } from 'react-icons/fi';
 import timestampConverter from '../../assets/utilities/timestampConverter';
 import getPath from '../../assets/utilities/getPath';
 import { PathContext } from '../../contexts/PathContext';
-import useAuth from '../../hooks/useAuth';
+import useAuth from '../../hooks/contexthooks/useAuth';
 import { api } from '../../api/resources';
 import styled from 'styled-components';
