+++ src/features/product/components/CatalogDetails.js
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import useAuth from '../../../hooks/useAuth';
+import useAuth from '../../../hooks/contexthooks/useAuth';
 import data from '../../../data/SEARCH.json';
 
 const CatalogDetails = ({ product }) => {
