+++ src/features/product/components/CatalogOwner.js
@@ -1,5 +1,5 @@
 import { Link } from 'react-router-dom';
-import useAuth from '../../../hooks/useAuth';
+import useAuth from '../../../hooks/contexthooks/useAuth';
 
 const CatalogOwner = ({ product }) => {
     const { isAuth } = useAuth();
