+++ src/features/search/Collection.js
@@ -10,8 +10,8 @@ import Form from '../../components/Form';
 import SearchInput from './components/SearchInput';
 import Button from '../../components/Button';
 import Loading from '../../layout/Loading';
-import useAuth from '../../hooks/useAuth';
-import useSearch from '../../hooks/useSearch';
+import useAuth from '../../hooks/contexthooks/useAuth';
+import useSearch from '../../hooks/contexthooks/useSearch';
 import { api } from '../../api/resources';
 import getViewPortWidth from '../../assets/utilities/getViewPortWidth';
 import hideSearchBar from '../../assets/utilities/hideSearchBar';
