+++ src/hooks/contexthooks/useCart.js
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { CartContext } from '../contexts/CartContext';
+import { CartContext } from '../../contexts/CartContext';
 
 const useCart = () => {
     return useContext(CartContext);
