+++ src/hooks/contexthooks/useMenu.js
@@ -1,0 +1,9 @@
+import { useContext } from 'react';
+import { MenuContext } from '../../contexts/MenuContext';
+
+const useMenu = () => {
+
+    return useContext(MenuContext);
+}
+
+export default useMenu
