--- src/hooks/useNav.js
@@ -1,9 +1,0 @@
-import { useContext } from 'react';
-import { NavContext } from '../contexts/NavContext';
-
-const useNav = () => {
-
-    return useContext(NavContext);
-}
-
-export default useNav
