+++ src/hooks/useNavIcons.js
@@ -1,0 +1,14 @@
+import React from 'react';
+import useMenu from './contexthooks/useMenu';
+
+const useNavIcons = () => {
+    const { displayMenu, setDisplayMenu } = useMenu();
+
+    return (
+        <div>
+
+        </div>
+    )
+}
+
+export default useNavIcons
