+++ src/layout/navigation/Hamburger.js
@@ -1,7 +1,7 @@
-import useNav from '../../hooks/useNav';
+import useMenu from '../../hooks/contexthooks/useMenu';
 
 const Hamburger = () => {
-    const { displayMenu, setDisplayMenu } = useNav();
+    const { displayMenu, setDisplayMenu } = useMenu();
 
     return (
         <label htmlFor="mobile-nav" className="hamburger mobile-nav-label" onClick={() => setDisplayMenu(!displayMenu)}>
