+++ src/layout/navigation/navIcons/SignIcon.js
@@ -2,6 +2,7 @@ import { Link } from 'react-router-dom';
 import { FaUserCircle } from 'react-icons/fa';
 
 const SignIcon = () => {
+
     return (
         <div id="signin-icon" className="nav-icon">
             <Link className="signin-icon" to='/login'>
