+++ src/pages/ShoppingCart.js
@@ -10,8 +10,8 @@ import Container from '../components/Container';
 import Aside from '../components/Aside';
 import Main from '../components/Main';
 import Count from '../features/search/components/Count';
-// import useAuth from '../hooks/useAuth';
-import useCart from '../hooks/useCart';
+// import useAuth from '../hooks/contexthooks/useAuth';
+import useCart from '../hooks/contexthooks/useCart';
 
 function ShoppingCart() {
   // Hooks
