+++ src/routes/PrivateRoutes.js
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import useAuth from '../hooks/useAuth';
+import useAuth from '../hooks/contexthooks/useAuth';
 
 // Requires authentication
 const PrivateRoutes = () => {
