+++ src/routes/PublicRoutes.js
@@ -1,5 +1,5 @@
 import { Outlet } from 'react-router-dom';
-import useAuth from '../hooks/useAuth';
+import useAuth from '../hooks/contexthooks/useAuth';
 
 // All access
 const PublicRoutes = () => {
