+++ src/routes/RestrictedRoutes.js
@@ -1,5 +1,5 @@
 import { Navigate, Outlet } from 'react-router-dom';
-import useAuth from '../hooks/useAuth';
+import useAuth from '../hooks/contexthooks/useAuth';
 
 // Can't be accessed if authenticated
 const RestrictedRoutes = () => {
