+++ src/styles/navbar.css
@@ -69,8 +69,7 @@ nav {
     pointer-events: none;
 }
 
-#auth-icon
-#signin-icon {
+#auth-icon {
     display: none;
 }
